Cleaning of code
Code was organized so that the sequence is more sensible. A game script was added that contains code that is applicable to both arcade and classic games. Comments were added to explain the code. Classic game: score is now hidden (only provided at the end of the game) and the algorithm to change the rule is changes so that the rule changed after 10 correct sorts. Arcade game: algorithm to change the rule is also changed, so that the rule changes after 3-5 correct sorts.  The algorithm was changed to ensure that a player masters the rule before it changes.